Debugged the issue where I was getting a float instead of an integer.

+++ python_introduction/finance_calculator.py
@@ -4,7 +4,7 @@ income = int(input("Enter your monthly income:"))
 expenses = int(input("Enter your total monthly expenses:"))
 
 monthly_savings = income - expenses
-annual_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
+annual_savings = int((monthly_savings * 12 + (monthly_savings * 12 * 0.05)))
 
 print("Your monthly savings are ${}.".format(monthly_savings))
 print(f"Projected savings after one year, with interest, is: ${annual_savings}.")
